add List of manipulation, add event's methodes

+++ src/main/java/clinic/dto/CaseDTO.java
@@ -6,13 +6,13 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.format.annotation.DateTimeFormat;
 
-import javax.persistence.*;
 import java.time.LocalDate;
+
 @Getter
 @Setter
 public class CaseDTO {
 
-    private Integer id;
+    private Long id;
 
     //patient
     private Patient patient;
@@ -34,11 +34,11 @@ public class CaseDTO {
     //status of the case (patient's therapy continues?)
     private boolean openCase;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
